Factor out location parsing in CLI
